js: bbox query fetches features in parallel batches
When fetching features from a bbox query, long ago we used to very
naively fetch the precise range of each feature. This was efficient in
bytes, but resulted in lots of requests.

Prior to this commit, whenever fetching new feature data, we'd
arbitrarily fetch some extra bytes on the chance that it'd obviate a
future request. In practice this performed fairly well - much better
than the naive approach. But you'd always have edge cases where you were
just a few bytes short of needed features, causing extra requests.

Rather than arbitrarily fetching an extra N bytes every time we fetch a
feature and hope it covers some soon-to-be-needed features, if we
instead knew the precise location and lengths of all the features we
need, we could explicitly combine the nearby ones into a single request.

Since the spatial index has the offsets of every feature*, we can do just
that - right-sizing our feature requests and surfacing a configuration
to let people tune the number of bytes they are willing to "spend" in
order to obviate a future request (see [src/ts/Config.ts]).

*the one exception is the very last feature. Since the "next" feature
serves as the bookend of "this* feature, we are left to just guess the
length of the last feature. In practice requesting a Range beyond the
end of the file results in fetching precisely what we need.