[rust] NBD: Extract Iter struct rather than track State param
The State parameter was introduced as a type check to make sure we
weren't calling methods at the wrong time - e.g. we can't iterate over a
reader before we've selected features.

Instead, we now extract all the state related it iteration into the
FeatureIter, achieving the same type checking in a more conventional way
(and with less code).

Also, the previous mechanism worked only for method calls, but not for
field access.  Some of the structs fields were only safe to access when
the state had a certain value - like feature_count only gets populated
*after* selection. This was already protected by the public method, but
now it's also not possible to access it via the field.