[rust] CI clears cache when Cargo.toml changes (#328)
Previously it was looking for Cargo.lock which we do not version
control.

Note that because we do not commit Cargo.lock, that this approach will
gradually pull in the latest dependencies as we rebuild.

If we want all dependencies to be built from cache, we'd need to commit
a Cargo.lock. The downside to that approach is we might miss a breakage
with a newer dependency release - one that users are likely to be using
and that we should probably know about.