[rust] Fix "UnexpectedEof" error when bbox results includes first item
Note: The error doesn't occur until reading the *next* item.

What's going on here:

We want to set `count` in the iter initializer.

FGB includes `count` in the header, but it's optional. Because
flatbuffer doesn't have null types, we can't distinguish between `count`
being unset and `count` truly being zero.

So the way we detect an "empty" FGB is to try to read from the feature
buffer in the iter's initializer. If the feature buffer is empty, this
read will fail and we can immediately mark the FGB reader as "finished",
which allows the iter to be well behaved elsewhere.

The other factor is that, as we read features, we track the offset into
the feature buffer (`cur_pos`) so that we know how far we need to Seek
from where we currently are when jumping to the next feature in the BBOX
results.

So, every time we read a feature_size, we've read four bytes and then
when we read the feature we've moved whatever the feature_len encoded in
those four bytes.

BUT THE THING IS - we weren't updating the cur_pos immediately after
reading feature_len, we updated it after reading the feature, and
included an extra 4 bytes at that point.

That usually works.

However - in the case of the very first feature, we've already added the
"4" to the feature offset. That's a sentinel value we check for to infer
it's the first feature so that we don't read the feature_len *again*.

So in the case that the bbox included the first feature, we were adding
an *extra* 4 to the cur_pos.

The fix is:

We now advance cur_pos immediately after we read the feature length.
After reading the feature, we only add feature_len, not feature_len+4.

I've also replaced the `finished` bool and the `cur_pos == 4` check
with a more explicable `state` enum.

Fixes https://github.com/flatgeobuf/flatgeobuf/issues/361