Fix too-small request sizing after making a large request.
Background:

To minimize the number of requests, nearby features are grouped into
batches which can be fetched in a single request. While processing the
first feature of the batch, we fetch all the data we need for the entire
batch.

However, because our HTTP library holds the entire request in memory, to
keep memory usage reasonable, there was logic to cap the max size of
that http request.

Bug:

If you ever hit that max http request size while fetching a feature
batch, each subsequent feature in that batch would require 2 requests
per feature — one to read the feature's length and another to read the
feature's data.

This logic would fix itself upon starting the subsequent batch,
until/unless another large request was encountered.

Fix:

We now calculate the size of the batch each request, rather than just
once per batch.

I also added a test client to test this and hopefully any other future
request related bugs in the future.